Remove the notion of "plain file" as it is confusing in a context which only directories and nothing but plain files.
Rationale: the notion "plain file" in particular irritates the
translation team, probably because in the Cloud FS context there are no
"special files" like sockets, pipes and the like. So we have only
folders and files, where "files" from the point of view of modern
operating systems are here always "plain files" as special files do not
exist in the cloud.